员工管理和分类管理两个大部分已经实现，实现了菜品页面的增加与修改

+++ src/main/java/com/wup/entity/Employee.java
@@ -20,7 +20,7 @@ public class Employee implements Serializable {
     /**
      * 主键
      */
-    private Integer id;
+    private Long id;
 
     /**
      * 姓名
@@ -70,11 +70,11 @@ public class Employee implements Serializable {
     /**
      * 创建人
      */
-    private Integer createUser;
+    private Long createUser;
 
     /**
      * 修改人
      */
-    private Integer updateUser;
+    private Long updateUser;
 
 }
