+++ src/main/java/com/wup/mapper/EmployeeMapper.java
@@ -16,7 +16,7 @@ import java.util.List;
 public interface EmployeeMapper {
     Employee findByUsername(Employee employee);
     Employee findByUsernameAndPassword(Employee employee);
-    Employee findById(Integer id);
+    Employee findById(Long id);
     void insert(Employee employee);
     List<Employee> getAllByName(String name);
     void update(Employee employee);
